% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iomDisparity.R
\name{iomDisparity.glm}
\alias{iomDisparity.glm}
\title{Rank and Replace method of adjustment for racial disparities}
\usage{
iomDisparity.glm(m, x, y, index, race, family = gaussian, link = "identity",
  ...)
}
\arguments{
\item{m}{a model.frame for the interior GLM function}

\item{x}{design matrix x, sans intercept}

\item{y}{target or dependent outcome}

\item{family}{generalized linear model family see help(glm), help(family), defaults to Gamma}

\item{link}{link function for generalized linear model}
}
\description{
Idea: fit multivariate generalized linear model y ~ X'b, collect coeffs b    
The rank-and-replace method adjusts health status by ranking each sample by  
a summary index of health status and replacing the health status of each     
minority individual with that of the correspondingly ranked white,           
thus preserving the ranking of health status and its rank correlation with   
SES measures.
}
\examples{
iomDisparity.glm()
}
\keyword{GLM,}
\keyword{disparities}
\keyword{disparities,}
\keyword{health}
\keyword{racial}
